1. Make the onedrive login can be run; 2. Update the test cases.
